Fixed problem with hparam plugin (?)
Tensorboard hparam plugin would ignore the hparam config record and just display all scalar values if the timestamp of the record was after the timestamp of the runs.
At least that's what I think the problem was. Until now the hparam plugin has just been full of bugs and unexpected behaviors.